:truck: [#3403] Break out models.py into separate module
The file is getting to large to manage/focus on the tasks at hand.
