+++ src/openforms/config/models/__init__.py
@@ -1,0 +1,9 @@
+from .color import RichTextColor
+from .config import GlobalConfiguration
+from .csp import CSPSetting
+
+__all__ = [
+    "CSPSetting",
+    "GlobalConfiguration",
+    "RichTextColor",
+]
