+++ src/openforms/config/models/config.py
@@ -1,10 +1,6 @@
 import logging
-from collections import defaultdict
 from functools import partial
 
-from django.contrib.admin.options import get_content_type_for_model
-from django.contrib.contenttypes.fields import GenericForeignKey
-from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ValidationError
 from django.core.validators import (
     FileExtensionValidator,
@@ -12,13 +8,11 @@ from django.core.validators import (
     MinValueValidator,
     RegexValidator,
 )
-from django.db import models, transaction
+from django.db import models
 from django.template.loader import render_to_string
 from django.utils.encoding import force_str
-from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 
-from colorfield.fields import ColorField
 from django_better_admin_arrayfield.models.fields import ArrayField
 from glom import glom
 from solo.models import SingletonModel
@@ -33,8 +27,8 @@ from openforms.translations.utils import ensure_default_language
 from openforms.utils.fields import SVGOrImageField
 from openforms.utils.translations import runtime_gettext
 
-from .constants import CSPDirective, UploadFileType
-from .utils import verify_clamav_connection
+from ..constants import UploadFileType
+from ..utils import verify_clamav_connection
 
 logger = logging.getLogger(__name__)
 
@@ -648,117 +642,3 @@ class GlobalConfiguration(SingletonModel):
 
     def get_theme_classname(self) -> str:
         return self.theme_classname or "openforms-theme"
-
-
-class CSPSettingQuerySet(models.QuerySet):
-    def as_dict(self):
-        ret = defaultdict(set)
-        for directive, value in self.values_list("directive", "value"):
-            ret[directive].add(value)
-        return {k: list(v) for k, v in ret.items()}
-
-
-class CSPSettingManager(models.Manager.from_queryset(CSPSettingQuerySet)):
-    @transaction.atomic
-    def set_for(
-        self,
-        obj: models.Model,
-        settings: list[tuple[CSPDirective, str]],
-        identifier: str = "",
-    ) -> None:
-        """
-        Deletes all the connected CSP settings and creates new ones based on the new provided data.
-
-        :param obj: The configuration model providing this CSP entry.
-        :param settings: A two-tuple containing values used to create the underlying ``CSPSetting`` model.
-        :param identifier: An optional string to further identify the source of this CSP entry.
-        """
-        instances = [
-            CSPSetting(
-                content_object=obj,
-                directive=directive,
-                value=value,
-                identifier=identifier,
-            )
-            for directive, value in settings
-        ]
-
-        CSPSetting.objects.filter(
-            content_type=get_content_type_for_model(obj),
-            object_id=str(obj.id),
-            identifier=identifier,
-        ).delete()
-
-        self.bulk_create(instances)
-
-
-class CSPSetting(models.Model):
-    directive = models.CharField(
-        _("directive"),
-        max_length=64,
-        choices=CSPDirective.choices,
-        help_text=_("CSP header directive."),
-    )
-    value = models.CharField(
-        _("value"),
-        max_length=255,
-        help_text=_("CSP header value."),
-    )
-
-    identifier = models.CharField(
-        _("identifier"),
-        max_length=64,
-        blank=True,
-        help_text=_("An extra tag for this CSP entry, to identify the exact source."),
-    )
-
-    # Generic relation fields (see https://docs.djangoproject.com/en/dev/ref/contrib/contenttypes/#generic-relations):
-    content_type = models.ForeignKey(
-        ContentType,
-        verbose_name=_("content type"),
-        on_delete=models.SET_NULL,
-        blank=True,
-        null=True,
-    )
-    object_id = models.TextField(
-        verbose_name=_("object id"),
-        blank=True,
-        db_index=True,
-    )
-    content_object = GenericForeignKey("content_type", "object_id")
-
-    objects: CSPSettingManager = CSPSettingManager()
-
-    class Meta:
-        ordering = ("directive", "value")
-
-    def __str__(self):
-        return f"{self.directive} '{self.value}'"
-
-
-class RichTextColor(models.Model):
-    color = ColorField(
-        _("color"),
-        format="hex",
-        help_text=_("Color in RGB hex format (#RRGGBB)"),
-    )
-    label = models.CharField(
-        _("label"),
-        max_length=64,
-        help_text=_("Human readable label for reference"),
-    )
-
-    class Meta:
-        verbose_name = _("text editor color preset")
-        verbose_name_plural = _("text editor color presets")
-        ordering = ("label",)
-
-    def __str__(self):
-        return f"{self.label} ({self.color})"
-
-    def example(self):
-        return mark_safe(
-            f'<span style="background-color: {self.color};">&nbsp; &nbsp; &nbsp;</span>'
-        )
-
-    example.short_description = _("Example")
